+++ document-ai-back/src/main/resources/db/migration/V2025_07_27_13_44_00__create_comment_and_evaluation_table.sql
@@ -7,7 +7,7 @@ create table comments (
 
 create table evaluations (
     id          uuid primary key default uuid_generate_v4(),
-    note        TINYINT not null,
+    note        smallint not null,
     created_at  timestamp default localtimestamp not null,
     document_id uuid not null references documents(id)
 );
